[TASK] Fix phpstan checkFunctionArgumentTypes errors in ext:install
This patch fixes incompatible type usage in function arguments
and is preparatory work for introducing native type hints and
strict mode in all core files.

Releases: master, 10.4
Resolves: #92108
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65598
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>

+++ typo3/sysext/install/Classes/FolderStructure/AbstractNode.php
@@ -187,7 +187,7 @@ abstract class AbstractNode
      */
     protected function getCurrentPermission()
     {
-        $permissions = decoct(fileperms($this->getAbsolutePath()));
+        $permissions = decoct((int)fileperms($this->getAbsolutePath()));
         return substr($permissions, -4);
     }
